Invented dataset with wages of men and women.

This dataset has been completely invented, in order to do some examples with the package.

invented_wages

Format

A data frame with 1000 rows and 5 variables:

gender

gender of the worker (men or women)

sector

economic sector where the worker is employed (secondary or tertiary)

education

educational level of the worker (I, II or III)

wage

monthly wage of the worker (in an invented currency)

sample_weights

sampling weights

Details

Every row of the dataset consists in a fake/invented individual worker. For every individual there is his/her gender, the economic sector in which he/she works, his/her level of education and his/her wage. Furthermore there is a column with the sampling weights.