A minimal function which counts the number of observations by groups in a data frame

A minimal function which counts the number of observations by groups in a data frame

jointfun_(.data, .variables, ...)

Arguments

.data
data frame to be processed

.variables
variables to split data frame by, as a character vector (c("var1", "var2")).

...
additional function calls to be applied on the .data

Value

a data frame, with a column for each cateogrical variable used, and a row for each combination of all the categorical variables' modalities.

Examples

data("invented_wages")
tmp <- jointfun_(.data = invented_wages, .variables = c("gender", "sector"))
tmp

#> # A tibble: 4 x 3
#> # Groups:   gender [2]
#>   gender sector        n
#>   <fct>  <fct>     <int>
#> 1 men    secondary   289
#> 2 men    tertiary    258
#> 3 women  secondary   166
#> 4 women  tertiary    287

str(tmp)

#> Classes ‘grouped_df’, ‘tbl_df’, ‘tbl’ and 'data.frame':	4 obs. of  3 variables:
#>  $ gender: Factor w/ 2 levels "men","women": 1 1 2 2
#>  $ sector: Factor w/ 2 levels "secondary","tertiary": 1 2 1 2
#>  $ n     : int  289 258 166 287
#>  - attr(*, "vars")= chr "gender"
#>  - attr(*, "drop")= logi TRUE
#>  - attr(*, "indices")=List of 2
#>   ..$ : int  0 1
#>   ..$ : int  2 3
#>  - attr(*, "group_sizes")= int  2 2
#>  - attr(*, "biggest_group_size")= int 2
#>  - attr(*, "labels")='data.frame':	2 obs. of  1 variable:
#>   ..$ gender: Factor w/ 2 levels "men","women": 1 2
#>   ..- attr(*, "vars")= chr "gender"
#>   ..- attr(*, "drop")= logi TRUE